Rename AppBar to SimpleAppBar to avoid conflict with material-ui.  Convert to Component.  Add Button for Login/Logout.
